Safari-611.1.10/LayoutTests/svg/custom/window-named-item-lookup.html

<!DOCTYPE html>
<html>
  <head>
    <script src="../../resources/js-test-pre.js"></script>
    <script>
      function runTest()
      {
        shouldBe("window.test1.length", "2");
        shouldBe("window.test2.length", "2");
        shouldBeEqualToString("window.test3.nodeName", 'A');
        shouldBe("window.test4.length", "2");
        shouldBe("window.test5.length", "2");
      }
     </script>
     <script src="../../resources/js-test-post.js"></script>
  </head>
  <body onload="runTest()">
    <a id="test1"></a>
    <svg id="test1"/>

    <svg id="test2"/>
    <a id="test2"></a>

    <a id="test3"></a>

    <a id="test4"></a>
    <div id="test4"></div>

    <svg id="test5"></svg>
    <svg id="test5"></svg>

  </body>
</html>
