third_party/blink/web_tests/svg/custom/window-named-item-lookup-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS window.test1.length is 2
PASS window.test2.length is 2
PASS window.test3.nodeName is "A"
PASS window.test4.length is 2
PASS window.test5.length is 2
 
 
